functions-api/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
azure-storage-blob==12.20.0
azure-cosmos==4.7.0
requests==2.32.3
python-dotenv==1.0.1
